Bring in the gh-pages example to simplify testing
Make it easier to get an example dialog while developing.

index.html and dialog.js are based on index.html from the gh-pages branch,
but the javascript has been split out and modified to work as a module.

Steps to run the test page:
- npm run build
- npm run start
(click the link to open the page)